main: Support the option to create an empty file or directory
A Node can create a new child for a file or a directory.
As it is now the dialog windows used to gather information
for the above process are a bit flickery.

There is no check to validate that the Node can have children
(aka its Inode is a Directory).